feat: Adding SignInPage and SignUpPage on viw path
