+++ web/index.html
@@ -33,6 +33,7 @@
 
     #app{
       display: flex;
+      flex-direction: column;
     }
   </style>
 </html>
